v1

+++ Intervalle.py
@@ -1,4 +1,4 @@
-nombre= int(input("Entrez un nombre décimal : "))
+nombre= float(input("Entrez un nombre décimal : "))
 x= nombre
 
 if x>=2 and x<3 or x>0 and x<=1 or x>=-10 and x<=-2:
